Prototyped support for a -import-dir option for Tosca issue #45.
For any non-system modules, Tosca will search the specified directories, defaulting to a directory that's relative to the main module.  The resulting URL and the base import directory that was used to find the module are both associated with the module during processing.  This is because the base is used in figuring out that Java package name and the Java or C++ file name for the generated file.